remove IsNotIntersection
