--- core/shared/src/main/scala-2/zio/IsNotIntersectionVersionSpecific.scala
@@ -1,6 +1,0 @@
-package zio
-
-trait IsNotIntersectionVersionSpecific {
-  implicit def materialize[A]: IsNotIntersection[A] =
-    macro zio.internal.macros.InternalMacros.materializeIsNotIntersection[A]
-}
