+++ core/shared/src/main/scala-3/zio/ServiceTagVersionSpecific.scala
@@ -15,13 +15,7 @@ private object TagMacros {
       case tpe if tpe.typeSymbol.isTypeParam || tpe.typeSymbol.isAbstractType =>
         Expr.summon[Tag[A]] match {
           case Some(tag) => tag
-          case None =>
-            (Expr.summon[EnvironmentTag[A]], Expr.summon[IsNotIntersection[A]]) match {
-              case (Some(tagExpr), Some(isNotIntersectionExpr)) =>
-              '{ Tag[A]($tagExpr, $isNotIntersectionExpr) }
-              case _ =>
-                report.errorAndAbort( s"Cannot find implicit Tag[${tpe.show}]" )
-            }
+          case None => report.errorAndAbort( s"Cannot find implicit Tag[${tpe.show}]" )
         }
       case AndType(_, _) =>
         report.errorAndAbort(s"You must not use an intersection type, yet have provided ${Type.show[A]}")
